// exercises-ts-belt/src/array/type.d.ts
export type Product = { name: string; price: number; inStock: boolean; };

type ProductExo22 = {
    name: string;
    price: number;
    category: string;
    quantity: number;
};

type ProductExo28 = {
    name: string;
    price: number;
    inStock: boolean;
    quantity: number;
};